Add a Fragment for the question.
