+++ app/src/main/res/layout/question_fragment_layout.xml
@@ -40,6 +40,5 @@
                 android:visibility="gone"
                 style="?android:attr/button"/>
         </LinearLayout>
-
 </LinearLayout>
